python-oop/m11-exam-preparation/Exam-Preparation-16-08-20/project/hardware/hardware.py
from project.software.software import Software


class Hardware:
    def __init__(self, name, type, capacity: int, memory: int) -> None:
        self.name = name
        self.type = type
        self.capacity = capacity
        self.memory = memory
        self.software_components = []

    def install(self, software: Software):
        if self.capacity >= software.capacity_consumption and self.memory >= software.memory_consumption:
            self.software_components.append(software)
        else:
            raise Exception('Software cannot be installed')

    def uninstall(self, software: Software):
        self.software_components.remove(software)
